Fixed bug when using asbint() with wc_get_page_id()
When a page does not exists `wc_get_page_id()` returns `-1`, so applying
`absint()` should cause bugs when comparing with "Hello world!" page.

+++ includes/wc-page-functions.php
@@ -154,7 +154,7 @@ function wc_nav_menu_item_classes( $menu_items ) {
 		return $menu_items;
 	}
 
-	$shop_page      = (int) wc_get_page_id( 'shop' );
+	$shop_page      = wc_get_page_id( 'shop' );
 	$page_for_posts = (int) get_option( 'page_for_posts' );
 
 	if ( ! empty( $menu_items ) && is_array( $menu_items ) ) {
